feat: add tool to retrieve todos for a specific date
